% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alraSeurat2.R
\name{alraSeurat2}
\alias{alraSeurat2}
\title{alraSeurat2}
\usage{
alraSeurat2(obj, assay.use = "RNA", slot.use = "data", ...)
}
\arguments{
\item{obj}{the Seurat object to run ALRA}

\item{assay.use}{The Seurat assay to impute. Default: "RNA"}

\item{slot.use}{The slot within the assay to impute. Default: "data"}

\item{...}{parameters passing to alra()}
}
\value{
an updated Seurat object with the ALRA imputed data in the "data" 
slot of the ALRA assay slot
}
\description{
Function to perform ALRA on a Seurat2 object
}
\details{
This function performs ALRA on the "data" slot in a given Seurat object,
does the neccessary pre- and post-processing to fit with the object, places 
the imputed data in an assay slot named "ALRA", and returns that updated Seurat 
object.
}
