% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b_nk_example}
\alias{b_nk_example}
\title{A matrix of 16427 rows and 12776 columns, mostly 0s.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 16427 rows and 12776 columns.
}
\usage{
b_nk_example
}
\description{
A matrix of 16427 rows and 12776 columns, mostly 0s.
}
\keyword{datasets}
