% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_k.R
\name{choose_k}
\alias{choose_k}
\title{choose_k}
\usage{
choose_k(A_norm, K = 100, pval_thresh = 1e-10, noise_start = 80,
  q = 2)
}
\arguments{
\item{A_norm}{The log-transformed expression matrix of cells (rows) vs. genes (columns)}

\item{K}{Number of singular values to compute. Must be less than the smallest dimension of the matrix.}

\item{pval_thresh}{The threshold for ``significance''}

\item{noise_start}{Index for which all smaller singular values are considered noise}

\item{q}{Number of additional power iterations}
}
\value{
A list with three items
}
\description{
Heuristic for choosing rank k for the low rank approximation based on
}
