% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labels_example}
\alias{labels_example}
\title{A label vector of containing 8344 "b_cells" and 8083 "cd56_nk"}
\format{
An object of class \code{factor} of length 16427.
}
\usage{
labels_example
}
\description{
A label vector of containing 8344 "b_cells" and 8083 "cd56_nk"
}
\keyword{datasets}
