% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alra.R
\name{normalize_data}
\alias{normalize_data}
\title{Simple convenience function to library and log normalize a matrix}
\usage{
normalize_data(A, verbose = TRUE)
}
\arguments{
\item{A}{A matrix of gene expression counts (genes x cells)}

\item{verbose}{Logical indicating whether to print progress messages}
}
\description{
Simple convenience function to library and log normalize a matrix
}
