% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alra.R
\name{alra}
\alias{alra}
\title{Computes the k-rank approximation to A_norm and adjusts it according to the
error distribution learned from the negative values.}
\usage{
alra(A_norm, k = 0, q = 10, quantile.prob = 0.001, use.mkl = F, mkl.seed = -1)
}
\arguments{
\item{A_norm}{The log-transformed expression matrix of cells (rows) vs. genes (columns)}

\item{k}{the rank of the rank-k approximation. Set to 0 for automated choice of k.}

\item{q}{the number of additional power iterations in randomized SVD}

\item{use.mkl}{Use the Intel MKL based implementation of SVD. Needs to be
installed from https://github.com/KlugerLab/rpca-mkl.}

\item{mkl.seed}{Only relevant if use.mkl=T. Set the seed for the random
generator for the Intel MKL implementation of SVD. Any number <0 will
use the current timestamp. If use.mkl=F, set the seed using
set.seed() function as usual.}
}
\value{
A list with three items
\enumerate{
\item The rank k approximation of A_norm.
\item The rank k approximation of A_norm, adaptively thresholded
\item The rank k approximation of A_norm, adaptively thresholded and
with the first two moments of the non-zero values matched to the
first two moments of the non-zeros of A_norm. This is the completed
matrix most people will want to work with
}
}
\description{
Computes the k-rank approximation to A_norm and adjusts it according to the
error distribution learned from the negative values.
}
\examples{
A_norm <- normalize_data(b_nk_example)
result.completed <- alra(A_norm,15)
A_norm_rank15 <- result.completed[[1]]     # The low rank approximation for reference purposes...not suggested for matrix completion
A_norm_rank15_cor <- result.completed[[3]] # The actual adjusted, completed matrix
}
