% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alra.R
\name{alra}
\alias{alra}
\title{alra}
\usage{
alra(A_norm, k = 0, q = 10, do.par = FALSE)
}
\arguments{
\item{A_norm}{The log-transformed expression matrix of cells (rows) vs. genes (columns)}

\item{k}{the rank of the rank-k approximation. Set to 0 for automated choice of k.}

\item{q}{the number of additional power iterations in randomized SVD}
}
\value{
A list with three items
}
\description{
Computes the k-rank approximation to A_norm and adjusts 
it according to the error distribution learned from the negative values.
}
\examples{
result.completed <- adjusted_svd(A_norm,15)
    A_norm_rank15 <- result.completed[[1]]     # The low rank approximation for reference purposes...not suggested for matrix completion
    A_norm_rank15_cor <- result.completed[[3]] # The actual adjusted, completed matrix
}
