% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alraSeurat2.R
\name{alraSeurat2}
\alias{alraSeurat2}
\title{Function to perform ALRA on a Seurat2 object}
\usage{
alraSeurat2(obj, ...)
}
\arguments{
\item{obj}{the Seurat object to run ALRA}

\item{...}{parameters passing to alra()}
}
\value{
an updated Seurat object with ALRA imputed data in the "data" slot
}
\description{
This function performs ALRA on the "data" slot in a given Seurat object,
does the neccessary pre- and post-processing to fit with the object, use
the ALRA imputed data to REPLACE the "data" slot in the Seurat object, and
returns that updated Seurat object.
}
