% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alra.R
\name{normalize_data}
\alias{normalize_data}
\title{Simple convenience function to library and log normalize a matrix}
\usage{
normalize_data(A)
}
\description{
Simple convenience function to library and log normalize a matrix
}
