% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{normalize_data}
\usage{
normalize_data(A)
}
\arguments{
\item{A}{matrix to be normalized}
}
\value{
a normalized matrix
}
\description{
Simple convenience function to library and log normalize a matrix
}
